feat: replace helidon-world with ws approach
