+++ helidon-world/src/main/java/io/helidon/examples/quickstart/se/Main.java
@@ -1,92 +1,108 @@
-/*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 
 package io.helidon.examples.quickstart.se;
 
 import java.io.IOException;
-import java.util.logging.LogManager;
+import java.net.URI;
 
+import javax.websocket.ClientEndpointConfig;
+import javax.websocket.DeploymentException;
+
+import org.glassfish.tyrus.client.ClientManager;
+
+import io.helidon.common.LogConfig;
+import io.helidon.common.reactive.Single;
 import io.helidon.config.Config;
+import io.helidon.health.HealthSupport;
+import io.helidon.health.checks.HealthChecks;
+import io.helidon.media.jsonp.JsonpSupport;
+import io.helidon.metrics.MetricsSupport;
 import io.helidon.webserver.Routing;
-import io.helidon.webserver.ServerConfiguration;
 import io.helidon.webserver.WebServer;
-import io.helidon.webserver.json.JsonSupport;
 
 /**
- * Simple Hello World rest application.
+ * The application main class.
  */
 public final class Main {
 
     /**
      * Cannot be instantiated.
      */
-    private Main() { }
-
-    /**
-     * Creates new {@link Routing}.
-     *
-     * @return the new instance
-     */
-    private static Routing createRouting() {
-        return Routing.builder()
-                .register(JsonSupport.get())
-                .register("/greet", new GreetService())
-                .register("/message", new MessageService())
-                .build();
+    private Main() {
     }
 
     /**
      * Application main entry point.
      * @param args command line arguments.
-     * @throws IOException if there are problems reading logging properties
      */
-    public static void main(final String[] args) throws IOException {
+    public static void main(final String[] args) {
+        startWebsocketClient();
         startServer();
+    }
+    
+    /**
+     * Start the websocket client
+     */
+    static void startWebsocketClient() {
+        ClientManager websocketClient = ClientManager.createClient();
+        URI websocketUri = URI.create("ws://localhost:11800/chat/helidon");
+        ClientEndpointConfig config = ClientEndpointConfig.Builder.create().build();
+        try {
+            websocketClient.connectToServer(new ChatClient(), config, websocketUri);
+        } catch (DeploymentException | IOException e) {
+            e.printStackTrace();
+        }
     }
 
     /**
      * Start the server.
      * @return the created {@link WebServer} instance
-     * @throws IOException if there are problems reading logging properties
      */
-    protected static WebServer startServer() throws IOException {
+    protected static Single<WebServer> startServer() {
 
         // load logging configuration
-        LogManager.getLogManager().readConfiguration(
-                Main.class.getResourceAsStream("/logging.properties"));
+        LogConfig.configureRuntime();
 
         // By default this will pick up application.yaml from the classpath
         Config config = Config.create();
 
-        // Get webserver config from the "server" section of application.yaml
-        ServerConfiguration serverConfig =
-                ServerConfiguration.fromConfig(config.get("server"));
+        WebServer server = WebServer.builder(createRouting(config))
+                .config(config.get("server"))
+                .addMediaSupport(JsonpSupport.create())
+                .build();
 
-        WebServer server = WebServer.create(serverConfig, createRouting());
+        Single<WebServer> webserver = server.start();
 
-        // Start the server and print some info.
-        server.start().thenAccept(ws -> {
-            System.out.println(
-                    "WEB server is up! http://localhost:" + ws.port());
-        });
+        // Try to start the server. If successful, print some info and arrange to
+        // print a message at shutdown. If unsuccessful, print the exception.
+        webserver.thenAccept(ws -> {
+                    System.out.println("WEB server is up! http://localhost:" + ws.port() + "/greet");
+                    ws.whenShutdown().thenRun(() -> System.out.println("WEB server is DOWN. Good bye!"));
+                })
+                .exceptionallyAccept(t -> {
+                    System.err.println("Startup failed: " + t.getMessage());
+                    t.printStackTrace(System.err);
+                });
 
-        // Server threads are not demon. NO need to block. Just react.
-        server.whenShutdown().thenRun(()
-                -> System.out.println("WEB server is DOWN. Good bye!"));
+        return webserver;
+    }
 
-        return server;
+    /**
+     * Creates new {@link Routing}.
+     *
+     * @return routing configured with JSON support, a health check, and a service
+     * @param config configuration of this server
+     */
+    private static Routing createRouting(Config config) {
+
+        MetricsSupport metrics = MetricsSupport.create();
+        HealthSupport health = HealthSupport.builder()
+                .addLiveness(HealthChecks.healthChecks())   // Adds a convenient set of checks
+                .build();
+
+        return Routing.builder()
+                .register(health)                   // Health at "/health"
+                .register(metrics)                  // Metrics at "/metrics"
+                .register("/message", new MessageService())
+                .build();
     }
 }
