+++ helidon-world/src/main/java/io/helidon/examples/quickstart/se/package-info.java
@@ -1,24 +1,2 @@
-/*
- * Copyright (c) 2017-2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 
-/**
- * Quickstart demo application
- * <p>
- * Start with {@link io.helidon.examples.quickstart.se.Main} class.
- *
- * @see io.helidon.examples.quickstart.se.Main
- */
 package io.helidon.examples.quickstart.se;
