+++ helidon-world/src/main/resources/logging.properties
@@ -1,35 +1,17 @@
-#
-# Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
 
 # Example Logging Configuration File
 # For more information see $JAVA_HOME/jre/lib/logging.properties
 
 # Send messages to the console
-handlers=java.util.logging.ConsoleHandler
+handlers=io.helidon.common.HelidonConsoleHandler
 
-# Global default logging level. Can be overriden by specific handlers and loggers
-.level=INFO
-
-# Helidon Web Server has a custom log formatter that extends SimpleFormatter.
-# It replaces "!thread!" with the current thread name
-java.util.logging.ConsoleHandler.level=INFO
-java.util.logging.ConsoleHandler.formatter=io.helidon.webserver.netty.WebServerLogFormatter
+# HelidonConsoleHandler uses a SimpleFormatter subclass that replaces "!thread!" with the current thread
 java.util.logging.SimpleFormatter.format=%1$tY.%1$tm.%1$td %1$tH:%1$tM:%1$tS %4$s %3$s !thread!: %5$s%6$s%n
 
-#Component specific log levels
+# Global logging level. Can be overridden by specific loggers
+.level=INFO
+
+# Component specific log levels
 #io.helidon.webserver.level=INFO
 #io.helidon.config.level=INFO
 #io.helidon.security.level=INFO
