+++ helidon-world/src/test/java/io/helidon/examples/quickstart/se/MainTest.java
@@ -37,7 +37,7 @@ public class MainTest {
 
     @BeforeAll
     public static void startTheServer() throws Exception {
-        webServer = Main.startServer();
+        webServer = Main.startServer().await();
         while (! webServer.isRunning()) {
             Thread.sleep(1 * 1000);
         }
@@ -50,34 +50,6 @@ public class MainTest {
                      .toCompletableFuture()
                      .get(10, TimeUnit.SECONDS);
         }
-    }
-
-    @Test
-    public void testHelloWorld() throws Exception {
-        HttpURLConnection conn;
-
-        conn = getURLConnection("GET","/greet");
-        Assertions.assertEquals(200, conn.getResponseCode(), "HTTP response1");
-        JsonReader jsonReader = Json.createReader(conn.getInputStream());
-        JsonObject jsonObject = jsonReader.readObject();
-        Assertions.assertEquals("Hello World!", jsonObject.getString("message"),
-                "default message");
-
-        conn = getURLConnection("GET", "/greet/Joe");
-        Assertions.assertEquals(200, conn.getResponseCode(), "HTTP response2");
-        jsonReader = Json.createReader(conn.getInputStream());
-        jsonObject = jsonReader.readObject();
-        Assertions.assertEquals("Hello Joe!", jsonObject.getString("message"),
-                "hello Joe message");
-
-        conn = getURLConnection("PUT", "/greet/greeting/Hola");
-        Assertions.assertEquals(200, conn.getResponseCode(), "HTTP response3");
-        conn = getURLConnection("GET", "/greet/Jose");
-        Assertions.assertEquals(200, conn.getResponseCode(), "HTTP response4");
-        jsonReader = Json.createReader(conn.getInputStream());
-        jsonObject = jsonReader.readObject();
-        Assertions.assertEquals("Hola Jose!", jsonObject.getString("message"),
-                "hola Jose message");
     }
 
     @Test
