--- helidon-ws-world/src/main/java/io/helidon/examples/quickstart/se/package-info.java
@@ -1,2 +1,0 @@
-
-package io.helidon.examples.quickstart.se;
