--- helidon-ws-world/src/main/resources/application.yaml
@@ -1,7 +1,0 @@
-
-app:
-  greeting: "Hello"
-
-server:
-  port: 8080
-  host: 0.0.0.0
